Rewrite conversion results to improve data returned to user
Previously processing a file only returned the name of the input and output files and a boolean result.

It turns out converting between formats is not cut and dry requiring more information to be returned to the user. This
change returns specific infomration about the reading and writiing process including custom text messages. The
information greatly improves a users ability to comprehend how the conversion went and what files might need to
be touched by hand after the conversion.