Replace use to user web pages with api endpoints
Data from each endpoint can be a 6th as small of the user web page.
This gives a significant speed boost as most of the time spent parsing tibiawiki is waiting for the data to download. It's also better for the free wiki as they don't have to serve as much data.